+++ src/Resource/MyEnvelope.php
@@ -9,17 +9,11 @@ use DateTime;
 class MyEnvelope extends BaseResource
 {
     public string $id;
-
     public string $status;
-
     public string $emailSubject;
-
     public string $emailBody;
-
     public ?string $senderName = null;
-
     public ?string $senderEmail = null;
-
     public ?DateTime $sentAt = null;
 
     /** @var Collection<MyEnvelopeDocument> */