+++ src/Resource/MyEnvelopeDocument.php
@@ -7,8 +7,6 @@ namespace DigitalCz\DigiSign\Resource;
 class MyEnvelopeDocument extends BaseResource
 {
     public string $id;
-
     public string $name;
-
     public string $fileMimeType;
 }