+++ src/Resource/EnvelopeTemplateRecipient.php
@@ -53,7 +53,7 @@ class EnvelopeTemplateRecipient extends BaseResource
 
     public ?string $address = null;
 
-    public ?SignatureScenarioVersionInfo $signatureScenarioVersionInfo = null;
+    public ?SignatureScenarioVersionInfo $scenarioVersionInfo = null;
 
     public ?IdentifyScenarioVersionInfo $identifyScenarioVersionInfo = null;
 }