+++ src/Resource/EnvelopeTemplateRecipient.php
@@ -52,4 +52,12 @@ class EnvelopeTemplateRecipient extends BaseResource
     public ?string $identificationNumber = null;
 
     public ?string $address = null;
+
+    public ?string $scenario = null;
+
+    public ?string $identifyScenario = null;
+
+    public ?SignatureScenarioVersionInfo $scenarioVersionInfo = null;
+
+    public ?IdentifyScenarioInfo $identifyScenarioInfo = null;
 }