+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [Unreleased]
+### Added
+- Add `AccountSecurity.auditLogDownloadDefaultUnchecked`
+
 ## [1.13.0] - 2023-05-09
 ### Added
 - Add `EnvelopeRecipientIdentification.authorizedAt`