+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `AccountSettings.identify`
 - Add multiple fields for `IdentifyScenarioVersion`
 - Add `IdentifyScenariosEndpoint.info` endpoint
+- Add `SignatureScenariosEndpoint.info` endpoint
 
 ### Changed
 - Change `Identification.envelope` from string to EnvelopeInfo