+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `IdentifyScenariosEndpoint.info` endpoint
 - Add `SignatureScenariosEndpoint.info` endpoint
 - Add scenario fields on `EnvelopeRecipient` and `EnvelopeTemplateRecipient`
+- Modify and add various `SignatureScenarioVariant` fields
 
 ### Changed
 - Change `Identification.envelope` from string to EnvelopeInfo